pseudo gen now according to shots

+++ news_sum/nsum_training.py
@@ -395,7 +395,7 @@ def TrainOneBatch(model, optimizer, scheduler, data, loss_fun, args):
     else:
         gradient = torch.ones((loss.shape[0]), dtype=torch.long)
 
-    loss.backward(gradient=gradient)
+    loss.backward(gradient=gradient.half())
     loss = loss.mean()
     optimizer.step()
     scheduler.step()
